Change fallback mode when no 'word' is specified
Falling back to the query led to confusing output: definitions that sometimes
didn't even match the search query, but were displayed as though they did.
Changed the code to just use the "readings" as the words in cases where only
reading(s) are defined, as is the case with a number of katakana entries.